Remove pagination to renderMarkup
