+++ src/index.js
@@ -1,4 +1,3 @@
-
 import('./js/teamModal');
 
 import './js/firebaseGoogleAuth';
@@ -7,4 +6,3 @@ import('./js/rendermarkup');
 
 import './js/apiFetches.js';
 import './js/test-fetchest.js';
-import './js/tui-pagination.js';
