+++ src/js/rendermarkup.js
@@ -1,11 +1,16 @@
 import NewApiFetches from './apiFetches';
 import Notiflix from 'notiflix';
+import Pagination from 'tui-pagination'; //add for pagination
+import 'tui-pagination/dist/tui-pagination.css'; //add for pagination
+
 const newApiFetches = new NewApiFetches();
 
 const refs = {
   gallery: document.querySelector('.film-cards'),
   searchForm: document.querySelector('.search-form'),
+  pagination: document.querySelector('.tui-pagination'), //add for pagination
 };
+let searchQuery = ''; //зробив глобальною, бо потрібна в файлі пагінації (add for pagination)
 
 const clearMarkup = () => {
   refs.gallery.innerHTML = '';
@@ -58,19 +63,70 @@ const handleSubmit = e => {
   e.preventDefault();
   const form = e.currentTarget;
   console.log(form.elements);
-  const searchQuery = form.elements.query.value.trim();
+  searchQuery = form.elements.query.value.trim();
 
   newApiFetches
-    .fetchSearchMovie(searchQuery)
+    .fetchSearchMovie(searchQuery, 1)
     .then(data => {
       if (data.length === 0) {
+        //  ----------------------add for pagination { -----------------------
+        const isHiddenClass = refs.pagination.classList.contains('is-hidden');
+        if (!isHiddenClass) {
+          refs.pagination.classList.add('is-hidden');
+        }
+        //  ----------------------} add for pagination -----------------------
         Notiflix.Notify.failure('Sorry! There is no movies with this title');
         return;
       }
       clearMarkup();
       renderMoviesListTemplate(data);
+
+      pagination.off('beforeMove', loadMoreFilms); //add for pagination
+      pagination.on('beforeMove', loadMoreFilms); //add for pagination
+      pagination.reset(); //add for pagination
+      refs.pagination.classList.remove('is-hidden'); //add for pagination
     })
     .catch(error => console.log(error));
 };
-
 refs.searchForm.addEventListener('submit', handleSubmit);
+
+//--------------------------- pagination option -----------------------------------
+const options = {
+  totalItems: 250,
+  itemsPerPage: 20,
+  visiblePages: 7,
+  page: 1,
+  centerAlign: true,
+  template: {
+    page: '<a href="#" class="tui-page-btn">{{page}}</a>',
+    currentPage:
+      '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
+    moveButton:
+      '<a href="#" class="tui-page-btn tui-{{type}} custom-class-{{type}}">' +
+      '<span class="tui-ico-{{type}}">{{type}}</span>' +
+      '</a>',
+    disabledMoveButton:
+      '<span class="tui-page-btn tui-is-disabled tui-{{type}} custom-class-{{type}}">' +
+      '<span class="tui-ico-{{type}}">{{type}}</span>' +
+      '</span>',
+    moreButton:
+      '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip custom-class-{{type}}">' +
+      '<span class="tui-ico-ellip">...</span>' +
+      '</a>',
+  },
+};
+
+const pagination = new Pagination(refs.pagination, options);
+
+function loadMoreFilms(e) {
+  const currentPage = e.page;
+
+  newApiFetches
+    .fetchSearchMovie(searchQuery, currentPage)
+    .then(data => {
+      renderMoviesListTemplate(data);
+      console.log('Search ', data);
+    })
+    .catch(error => console.log(error));
+}
+//--------------------------- /pagination option -----------------------------------
