+++ src/page_home/body.html
@@ -3,3 +3,4 @@
     <ul class="film-cards"></ul>
   </div>
 </div>
+<include src="./page_home/tui-pagination.html"></include> <!-- add for pagination -->
